fix(lint): split pure helpers out of screen-helpers.tsx
react-refresh/only-export-components requires a component file to export
only components. screen-helpers.tsx mixed 5 pure helper functions with
its components, failing the Lint CI job. Move the helpers to a new
screen-format.ts and import them where needed.
